refactor: put everything in octoliner package
